project2-all-tests-passed-without-customExceptions

+++ src/main/java/com/example/repository/MessageRepository.java
@@ -6,6 +6,6 @@ import java.util.List;
 import org.springframework.data.jpa.repository.JpaRepository;
 import com.example.entity.Message;
 
-public interface MessageRepository extends JpaRepository<Message, Long>{
-    List<Message> findAllByPostedBy(Long postedBy);
+public interface MessageRepository extends JpaRepository<Message, Integer>{
+    List<Message> findAllByPostedBy(int postedBy);
 }
